Source cleanups, fixes, optimizations, comments.
git-svn-id: https://zxtune.googlecode.com/svn/trunk@431 e19a5227-2556-0410-816f-a35683127991

+++ src/core/plugins/players/stp_supp.cpp
@@ -798,7 +798,7 @@ namespace
         }
         else
         {
-          chunk.Data[AYM::DataChunk::REG_TONEN] = curSampleLine.Noise;
+          chunk.Data[AYM::DataChunk::REG_TONEN] = static_cast<uint8_t>(curSampleLine.Noise);
           chunk.Mask |= 1 << AYM::DataChunk::REG_TONEN;
         }
